[Build] Fix compile warnings across Audio, Core, Display, SVG, Tiri, and Vector modules
- Replace deprecated F2I() with std::lrint() throughout (audio, core, display, svg, vector)
- Add [[maybe_unused]] to platform-guarded static functions in audio and display
- Add static to dpms_name() definition to match its forward declaration
- Fix sign-compare warnings in SVG by casting m->Tag to unsigned int
- Suppress trigraph warning in Tiri IR emitter by splitting the string literal
- Convert #warning pragma in audio to a comment

+++ src/svg/anim_timing.cpp
@@ -154,7 +154,7 @@ static ERR animation_timer(extSVG *SVG, int64_t TimeElapsed, int64_t CurrentTime
       VectorMatrix *m = nullptr;
       if (vt.transforms.front()->additive IS ADD::REPLACE) {
          for (m = vector->Matrices; (m); m=m->Next) {
-            if (m->Tag IS MTAG_SVG_TRANSFORM) {
+            if ((unsigned int)m->Tag IS MTAG_SVG_TRANSFORM) {
                double d = 1.0 / (m->ScaleX * m->ScaleY - m->ShearY * m->ShearX);
 
                double t0 = m->ScaleY * d;
